Data.Strict.Tuple

Description

The strict variant of the standard Haskell pairs and the corresponding variants of the functions from Data.Tuple.

Note that unlike regular Haskell pairs, (x :!: _|_) = (_|_ :!: y) = _|_

Documentation

data Pair a b

The type of strict pairs.

Constructors

!a :!: !b infix 2

type (:!:) = Pair infix 2

fst :: Pair a b -> a

Extract the first component of a strict pair.

snd :: Pair a b -> b

Extract the second component of a strict pair.

curry :: (Pair a b -> c) -> a -> b -> c

Curry a function on strict pairs.

uncurry :: (a -> b -> c) -> Pair a b -> c

Convert a curried function to a function on strict pairs.

swap :: Pair a b -> Pair b a

Analogous to swap from Data.Tuple

zip :: [a] -> [b] -> [Pair a b]

Zip for strict pairs (defined with zipWith).

unzip :: [Pair a b] -> ([a], [b])

Unzip for stict pairs into a (lazy) pair of lists.