Data.Strict.Either

Description

Strict Either.

Same as the standard Haskell Either, but Left _|_ = Right _|_ = _|_

Documentation

data Either a b

The strict choice type.

Constructors

Left !a

Right !b

either :: (a -> c) -> (b -> c) -> Either a b -> c

Case analysis: if the value is Left a, apply the first function to a; if it is Right b, apply the second function to b.

isLeft :: Either a b -> Bool

Yields True iff the argument is of the form Left _.

isRight :: Either a b -> Bool

Yields True iff the argument is of the form Right _.

fromLeft :: Either a b -> a

Extracts the element out of a Left and throws an error if the argument is a Right.

fromRight :: Either a b -> b

Extracts the element out of a Right and throws an error if the argument is a Left.